// src/modules/cash-flow/dtos/controls/me-prices-controls.dto.ts
export class MePricesControlsDto {
  dt: {
    pAcem: number;
    pPeito: number;
    pGorduraExt: number;
    pGorduraInt: number;
    pMusculo: number;
    pPaleta: number;
  };
  pa: { pCostela: number };
  tr: {
    pBananinha: number;
    pLagarto: number;
    pContraFile: number;
    pMusculoMole: number;
    pCoxaoDuro: number;
    pMusculoDuro: number;
    pCoxaoMole: number;
    pPatinho: number;
    pFileMignon: number;
    pRecortes: number;
    pFileCostela: number;
    pRoubado: number;
    pCorAlcatra: number;
    pMaminha: number;
    pPicanha: number;
    pFralda: number;
  };
}